New config option replace_placeholders
Making the replacement of placeholders with their values configurable.

The previous issue where placeholders weren't replaced with the values
if they were set by bind_params() (fixed by 75279d3) was actually handy
for one use I made of this, where I wanted to parse the log and get a
count of how many times each individual query ran, ignoring the values
used for each execution - so the ability to disable replacement of the
placeholders seems handy.