Set autoflush on output filehandle
If we're outputting to a file, set autoflush on the handle, so that each
query we log ends up in the file right away; otherwise, with the default
block-buffered output, the query might not appear in the log until some
time later, which can be confusing.

It's possible that this could carry a mild performance penalty, but, if
you're logging all DB queries to a file, hopefully you're doing that
temporarily in a dev environment, not in production!